Crate simple_calculator

Enums

ReturnCodes

Constants

ARGS_LIST_MSG

DIV_CHAR

EXIT_INPUT

EXIT_MSG

HELP_ARG

HELP_ARG_MSG

H_ARG

MAIN_ERROR_MSG

MUL_CHAR

POW_STR

SUB_CHAR

SUM_CHAR

SYNTAX_ERROR_MSG

UNKNOWN_ARG_MSG

Functions

calculate
Does the calculation process.
@param expression_vector String vector with all the expressions.
@param buffer for storing the result.
@return ReturnCode depending the succes of the operation.

input
User input.
@param buffer for store the input.
@return ReturnCode depending the succes of the operation.

parse
Prepare the user input to be passed to the calculator. Clean spaces and does a syntax check.
@param input user input in String format.
@param String buffer for storing the parsed String.
@return ReturnCode depending the succes of the operation.

vectorize
Split the input expression (parsed) into a vector of expressions. Each expression is separated from the other by '+' or '-'.
@param expression_str input String parsed.
@param buffer_vector buffer for storing the vectorized expression.
@return ReturnCode depending the succes of the operation.